dev

See merge request minet/development/websites/minet/adh6.minet.net!99

+++ backend/adh6/treasury/storage/transaction_repository.py
@@ -1,3 +1,5 @@
+# pyright: reportIncompatibleMethodOverride=false
+
 """
 Implements everything related to actions on the SQL database.
 """
@@ -8,11 +10,12 @@ from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from adh6.constants import DEFAULT_LIMIT, DEFAULT_OFFSET
+from adh6.datetime_utils import utc_now_naive
 from adh6.entity import AbstractTransaction, Transaction
 from adh6.exceptions import PaymentMethodNotFoundError
 from adh6.storage.count import count_rows
+from adh6.treasury.interfaces import TransactionRepository
 
-from ..interfaces import TransactionRepository
 from .models import PaymentMethod, Transaction as SQLTransaction
 
 
@@ -68,8 +71,8 @@ class TransactionSQLRepository(TransactionRepository):
 
         return [_map_transaction_sql_to_entity(i) for i in r], count
 
-    async def create(self, abstract_transaction: AbstractTransaction) -> object:
-        now = datetime.now()
+    async def create(self, abstract_transaction: AbstractTransaction) -> Transaction:
+        now = utc_now_naive()
 
         method_id = None
         if abstract_transaction.payment_method is not None:
@@ -96,14 +99,15 @@ class TransactionSQLRepository(TransactionRepository):
 
         return _map_transaction_sql_to_entity(transaction)
 
-    def update(self, abstract_transaction: AbstractTransaction, override=False) -> object:
+    async def update(self, abstract_transaction: AbstractTransaction, override: bool = False) -> Transaction:
         raise NotImplementedError
 
-    async def delete(self, object_id) -> None:
+    async def delete(self, object_id: int) -> None:
         stmt = select(SQLTransaction).where(SQLTransaction.id == object_id)
         transaction = await self.session.scalar(stmt)
 
-        await self.session.delete(transaction)
+        if transaction is not None:
+            await self.session.delete(transaction)
 
     async def search_for_export(
         self,
